Add deprecation warning for HdfsClient
