+++ python/pyarrow/util.py
@@ -15,6 +15,8 @@
 # specific language governing permissions and limitations
 # under the License.
 
+import warnings
+
 # Miscellaneous utility code
 
 
@@ -23,3 +25,13 @@ def implements(f):
         g.__doc__ = f.__doc__
         return g
     return decorator
+
+
+def _deprecate_class(old_name, new_name, klass, next_version='0.5.0'):
+    msg = ('pyarrow.{0} is deprecated as of {1}, please use {2} instead'
+           .format(old_name, next_version, new_name))
+
+    def deprecated_factory(*args, **kwargs):
+        warnings.warn(msg, FutureWarning)
+        return klass(*args)
+    return deprecated_factory
